Ss 67 club edit clubinfo

+++ src/main/java/org/example/backendfootvolley/repository/NationalFederationRepository.java
@@ -3,5 +3,5 @@ package org.example.backendfootvolley.repository;
 import org.example.backendfootvolley.model.NationalFederation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NationalFederationRepository extends JpaRepository<NationalFederation, Long> {
+public interface NationalFederationRepository extends JpaRepository<NationalFederation, Integer> {
 }
